refactor: 이동 로직 interface를 이용하여 구현
