+++ src/main/java/step6/domain/Cars.java
@@ -3,18 +3,11 @@ package step6.domain;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
-import java.util.Random;
 import java.util.stream.Collectors;
 
 public class Cars {
     private static final String CAR_NAMES_DELIMITER = ",";
-    private static final int MAX_RANDOM_NUMBER = 9;
-    private static final Random rd = new Random();
     private final List<Car> cars;
-
-    protected Cars(List<Car> cars) {
-        this.cars = cars;
-    }
 
     public Cars(String inputCarName) {
         this.cars = validCarNames(Arrays.asList(inputCarName.split(CAR_NAMES_DELIMITER)));
@@ -28,11 +21,7 @@ public class Cars {
 
     public void runRacingGame(TryNumber tryNumber) {
         cars.stream()
-                .forEach(car -> car.move(getRandomNumber()));
-    }
-
-    private int getRandomNumber() {
-        return rd.nextInt(MAX_RANDOM_NUMBER) + 1;
+                .forEach(car -> car.move(new RandomMoveStrategy()));
     }
 
     public List<Car> getCars() {
