+++ src/main/java/step6/domain/MoveStrategy.java
@@ -1,0 +1,5 @@
+package step6.domain;
+
+public interface MoveStrategy {
+    boolean move();
+}
